# pkg/trace: optimize ExtractSubtraces