--- a/pkg/trace/stats/subtrace_test.go
+++ b/pkg/trace/stats/subtrace_test.go
@@ -137,3 +137,51 @@ func TestExtractSubtracesMeasuredSpans(t *testing.T) {
 	}
 
 }
+
+func BenchmarkExtractSubtraceSmall(b *testing.B) {
+	trace := pb.Trace{
+		&pb.Span{SpanID: 1, ParentID: 0, Service: "s1"},
+		&pb.Span{SpanID: 2, ParentID: 1, Service: "s2"},
+		&pb.Span{SpanID: 3, ParentID: 2, Service: "s2"},
+		&pb.Span{SpanID: 4, ParentID: 3, Service: "s2"},
+		&pb.Span{SpanID: 5, ParentID: 1, Service: "s1"},
+	}
+
+	traceutil.ComputeTopLevel(trace)
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		ExtractSubtraces(trace, trace[0])
+	}
+}
+
+func BenchmarkExtractSubtraceLarge(b *testing.B) {
+	trace := pb.Trace{
+		&pb.Span{SpanID: 1, ParentID: 0, Service: "s1"},
+		&pb.Span{SpanID: 2, ParentID: 1, Service: "s2"},
+		&pb.Span{SpanID: 3, ParentID: 2, Service: "s2"},
+		&pb.Span{SpanID: 4, ParentID: 3, Service: "s2"},
+		&pb.Span{SpanID: 5, ParentID: 1, Service: "s1"},
+		&pb.Span{SpanID: 6, ParentID: 1, Service: "s3"},
+		&pb.Span{SpanID: 7, ParentID: 1, Service: "s4"},
+	}
+
+	nextId := trace[len(trace)-1].SpanID + 1
+
+	for i := 0; i < 1000; i++ {
+		trace = append(trace, &pb.Span{SpanID: nextId, ParentID: 6, Service: "s3"})
+		nextId++
+	}
+
+	for i := 0; i < 1000; i++ {
+		trace = append(trace, &pb.Span{SpanID: nextId, ParentID: 7, Service: "s4"})
+		nextId++
+	}
+
+	traceutil.ComputeTopLevel(trace)
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		ExtractSubtraces(trace, trace[0])
+	}
+}