### What does this PR do?

This PR makes several improvements to the ExtractSubtraces:
- use an array instead of a linked list for the stack to avoid heap allocations
- remove useless operations from ChildrenMap
- early return for single span traces

```
name                    old time/op    new time/op    delta
ExtractSubtraceSmall-4    2.03µs ± 3%    1.56µs ± 1%  -23.21%  (p=0.008 n=5+5)
ExtractSubtraceLarge-4     890µs ± 4%     346µs ± 1%  -61.08%  (p=0.008 n=5+5)

name                    old alloc/op   new alloc/op   delta
ExtractSubtraceSmall-4      912B ± 0%      800B ± 0%  -12.28%  (p=0.008 n=5+5)
ExtractSubtraceLarge-4     595kB ± 0%     239kB ± 0%  -59.75%  (p=0.008 n=5+5)

name                    old allocs/op  new allocs/op  delta
ExtractSubtraceSmall-4      27.0 ± 0%      17.0 ± 0%  -37.04%  (p=0.008 n=5+5)
ExtractSubtraceLarge-4     4.13k ± 0%     0.08k ± 0%  -98.16%  (p=0.008 n=5+5)
```

### Motivation

Improve performance

### Additional Notes

Anything else we should know when reviewing?

### Describe your test plan

Write there any instructions and details you may have to test your PR.